requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.10.0
requests==2.32.0
python-dotenv==1.0.1
openai==1.58.1
langchain==0.3.12
langchain-openai==0.2.14
streamlit==1.41.0
httpx==0.27.0
pytest==8.3.4
